add useEditorSyncLabels util
